modules/data-pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-data-pipeline"
version = "0.1.0"
description = "Polygon.io download helpers and sync scripts"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Trading Systems" }]
dependencies = [
  "pandas>=2.2",
  "polygon-api-client>=1.15",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "google-cloud-bigquery>=3.25",
  "pyarrow>=23.0.1"
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov>=4"]

[project.scripts]
trading-data-download = "trading_data_pipeline.cli:main"
trading-data-sync = "trading_data_pipeline.bigquery_sync:main"
trading-data-pull = "trading_data_pipeline.bigquery_pull:main"
trading-data-visualize = "trading_data_pipeline.visualize:main"
trading-data-download-cmc = "trading_data_pipeline.coinmarketcap_downloader:main"

[tool.hatch.build.targets.wheel]
packages = ["src/trading_data_pipeline"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
